+++ app/src/client/client.cpp
@@ -9,11 +9,11 @@ std::shared_ptr<Client> create(const AppRemote& c) {
     std::string scheme = c.url.substr(0, pos);
     if (scheme == "webdav") {
         std::string url = "http" + c.url.substr(pos);
-        return std::make_shared<Webdav>(url, c.user, c.passwd);
+        return std::make_shared<Webdav>(url, c);
     }
     if (scheme == "webdavs") {
         std::string url = "https" + c.url.substr(pos);
-        return std::make_shared<Webdav>(url, c.user, c.passwd);
+        return std::make_shared<Webdav>(url, c);
     }
     if (scheme == "file") {
         return std::make_shared<Local>(c.url.substr(pos + 3));